requirements.txt
flask
flask-cors
adafruit-circuitpython-ads1x15
rpi-lgpio
torch
pandas